switch_item moved to fusion.lib as it's only place where is used
