+++ pype/hosts/fusion/scripts/fusion_switch_shot.py
@@ -234,7 +234,7 @@ def switch(asset_name, filepath=None, new=True):
     representations = []
     for container in containers:
         try:
-            representation = pype.switch_item(
+            representation = fusion_lib.switch_item(
                 container,
                 asset_name=asset_name)
             representations.append(representation)
